lib:CPUID:Digest: Implemented Zen2,Zen3,Zen4 detection
Now Zen2, Zen3, Zen4 can be checked.

Sha3: Updated CPUID calls for it.

It's better to current algorithm from arch/zen3 to avx2 as there is no
machine specific instructions used inside it. A copy of the same code in
arch/avx2 might also be enough to yield better speeds.

Signed-off-by: Abhiram S <abhiram.s@amd.com>