Scripts:Python: Prevent Malformed Copyright.
Copyrights which are not correctly formed, files without copyrights,
files without correct start/end year will be prevented from being
committed to the repo.

Also added git hook.

This script yet does not ensure the Copyright is staged for the current
commit (Partial Staging). - Will need to work on it.

Signed-off-by: Abhiram S <abhiram.s@amd.com>

Co-authored-by: Abhijith N Raj <abhijith.raj@amd.com>